NCGNT-708: Fix Regression: I can no longer delete my own flashcards
Found some more places where we had a type mismatch for user ids.
Updated the code to use number type for `user_id` which comes directly
from the API and then use string for `userId` which exists on our client
model data.

Added some more API based model classes so we have more type information
for what the API is giving us.

Updated the flashcards to convert the ids to string before comparison so
that the isOwn evaluates correctly which allows the delete action to
appear in the UI so the user can delete their flashcards.