CWEB-7: Streamline pull request template for `circlein-web` project
Added a precommit command that runs the unit tests and linter.  Also
added a precommit:fix that will run the linter fix and the tests.  The
fix command should only be used before the final commit so that any
fixed changes are committed.

Updated pull request template to make it fast to fill out (no need to
delete duplicated instructional text) and updated checklist items.