CWEB-168: Move hud story to the top of the page (underneath navbar)
Update the headers of all pages to have a consistent header that is compatible with the story conversation.

Renamed item.tsx so the component matches the filename.

Fixed issue where create based routes in the feed (/create_post) were
not handling the routing correctly resulting in a mismatch between the
URL and navigation areas.  Now they are consistent.

Converted store to be a functional component so that we can cleanly use
the story hooks with it.

Reorder navbar study tools to match right side order.