CWEB-283: Add type definitions for chat state
Added type definitions in chat state in reducers/chat.  Update any types
to be only used when we need to access a less-typed or private state
within the chat code.

Also fix an issue that adding types uncovered: If you add users to your
chat and you search for users, we don’t properly filter out the members
already in the channel. This was because we were filtering based on a
nonexistent 'id' property instead of the correct 'userId' property.